Move instance file
