+++ l10n.yaml
@@ -4,6 +4,5 @@ template-arb-file: intl_en.arb
 output-localization-file: l10n.dart
 preferred-supported-locales: [en]
 output-class: L10n
-untranslated-messages-file: assets/l10n/untranslated.json
 synthetic-package: false
 nullable-getter: false
